feat: flutter 노트와 통신 활성화 다만 파라미터로 받아지질 않음;;

+++ note_crud/models.py
@@ -5,5 +5,5 @@ from django.db import models
 class Note(models.Model):
     title = models.CharField(max_length=100)
     content = models.TextField()
-    hexColor = models.IntegerField()
+    hexColor = models.BigIntegerField()
     datetime = models.DateTimeField(auto_now=True)
